requirements.txt
pymodbus==3.7.4
aiomqtt==2.3.0
pyyaml==6.0.2
hypothesis==6.115.6
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.8.4
